change copy in remote services settings
